refactor: Delivery saving
